complete homework
judge if the graph is singly connected by finding forward of cross
edges. When we do dfs searching, if we discover a edge who leads to a
black vertex, we find a forward or cross edge, which means that the
current graph is not singly connected. however, we still have some
trouble of test data 17